audio: implemented max concurrent audio samples limit to reduce audio queue size, added placeholder for 3D emitter code
