+++ game/Content/Audio/AudioConfig.yaml
@@ -70,15 +70,13 @@ Category:
       Id: Female_02_loop
       Volume: 0.25
       Sound: "RacialMusic/Female_02_loop.m4a"
-    - SoundEffect:
-      Id: "Stardrive_Combat 1c_114BPM"
-      Volume: 0.25
-      Sound: "RacialMusic/Stardrive_Combat 1c_114BPM.m4a"
 
 Category:
   Name: Default
   Volume: 1.00
   FadeOutTime: 0.0 # The warp sounds shouldn't fade out because they are always one-shot anyways
+  MemoryCache: true # Sound effects in this category can be cached to memory to speed up the audio playback
+  MaxConcurrentSounds: 32 # Maximum number of sounds that can be played at once in this category
   SoundEffects:
     - SoundEffect:
       Id: sd_battle_ambient
@@ -144,6 +142,8 @@ Category:
 Category:
   Name: Weapons
   Volume: 1.00
+  MemoryCache: true # Sound effects in this category can be cached to memory to speed up the audio playback
+  MaxConcurrentSounds: 64 # The maximum number of sounds that can be played at the same time, 0 for unlimited
   SoundEffects:
     - SoundEffect:
       Id: alarm
