--- game/Content/Audio/RacialMusic/Stardrive_Combat 1c_114BPM.m4a
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:48de81a987ff5f9a7e9c520f9a01d92e2bd52419aa63787195e77d3c13ff3630
-size 1838910
